New feature: use SVM classifier
Replace KNN classifier BY Svm classifier. This classifier will lead to make a better
results than KNN.As shown in the result model accuracy and F_1 Score.
The is ofcourse a trade of between using SVM insted of KNN in this problem. Because SVM git
good results but it take large time to evaluate